+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SQLScriptQueries` Composite to run one or more SQL Scripts against a SQL Server/Instance
 - `RemoteDesktopCertificates` Composite to import Remote Desktop Certificates. Ideally combined with CertificateRequests and CertificateExports composites.
 - `RemoteDesktopHAMode` Composite to configure High Availability mode on a RDS connection broker.
+- `AzureConnectedMachine` Composite to install and configure the Azure Connected Machine Agent
 
 ### Changed